Backend/requirements.txt
autogen-agentchat
autogen-ext[openai]
python-dotenv
flask
flask-cors
